Add way to filter out specific sites
Client-Side rendered Next.js applications will return a valid JSON for `document.body.textContent`,
since they're preconfigured with that stuff.
This will trigger this extension and thus break those kinds of sites.

You can reproduce this behaviour by going to https://demo.ente.app with an earlier version of this extension.

This commit introduces a way to check sites for "red flags", which are basically just predicates that prevent a page from being picked up by this extension.
For Next.js apps, this this is the existance of `div#__next`.